refactor: move DevBenchBridge out of /src
